backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.0
openai==1.55.3
httpx==0.27.2          # bardzo ważne — nowe wersje łamią openai client
python-dotenv==1.0.1
pypdf==4.2.0
duckdb==1.0.0
pyarrow==16.1.0
numpy==1.26.4          # wymagane przez chroma/pyarrow
pydantic==2.7.4        # stabilne z fastapi
